use env python inside ansible
This ensures that both the python executing ansible and the python
called by ansible in tasks are the same, and see the same installed
packages, fixing a bug where ansible could not find the ``pexpect``
package.

Also use ``become: yes`` when doing the ``chsh`` to ``fish``,
ensuring that the user does not need to type the password at that
step. #1